# [libc++][format][3/7] Improves std::format performance.